+++ marketplace/services/vtex/generic_service.py
@@ -1,7 +1,13 @@
 """
 Service for managing VTEX App instances within a project.
 """
+
+import time
+
 from datetime import datetime
+
+from django.db import close_old_connections
+from django_redis import get_redis_connection
 
 from dataclasses import dataclass
 
@@ -41,7 +47,7 @@ class APICredentials:
         }
 
 
-class VtexService:
+class VtexServiceBase:
     fb_service_class = FacebookService
     fb_client_class = FacebookClient
 
@@ -102,6 +108,8 @@ class VtexService:
 
         return domain, app_key, app_token
 
+
+class ProductInsertionService(VtexServiceBase):
     def first_product_insert(self, credentials: APICredentials, catalog: Catalog):
         pvt_service = self.get_private_service(
             credentials.app_key, credentials.app_token
@@ -113,34 +121,23 @@ class VtexService:
             return None
 
         products_csv = pvt_service.data_processor.products_to_csv(products)
-        product_feed = self._send_products_to_facebook(products_csv, catalog)
-        self.product_manager.create_or_update_products_on_database(
-            products, catalog, product_feed
-        )
+        close_old_connections()
+        self._send_products_to_facebook(products_csv, catalog)
+        pvt_service.data_processor.clear_csv_buffer(
+            products_csv
+        )  # frees the memory of the csv file
+        # Removed on 03-30-2024
+        # close_old_connections()
+        # self.product_manager.create_or_update_products_on_database(
+        #     products, catalog, product_feed
+        # )
+        close_old_connections()
         self.app_manager.initial_sync_products_completed(catalog.vtex_app)
 
-        return pvt_service.data_processor.convert_dtos_to_dicts_list(products)
-
-    def webhook_product_insert(
-        self, credentials: APICredentials, catalog: Catalog, webhook_data, product_feed
-    ):
-        pvt_service = self.get_private_service(
-            credentials.app_key, credentials.app_token
-        )
-        products_dto = pvt_service.update_webhook_product_info(
-            credentials.domain, webhook_data, catalog.vtex_app.config
-        )
-        if not products_dto:
-            return None
-
-        products_csv = pvt_service.data_processor.products_to_csv(products_dto)
-        self._update_products_on_facebook(products_csv, catalog, product_feed)
-        self.product_manager.create_or_update_products_on_database(
-            products_dto, catalog, product_feed
-        )
-        return pvt_service.data_processor.convert_dtos_to_dicts_list(products_dto)
+        return products
 
     def _send_products_to_facebook(self, products_csv, catalog: Catalog):
+        print("Starting to upload the CSV file to Facebook")
         current_time = datetime.now().strftime("%Y-%m-%d_%H-%M")
         file_name = f"csv_vtex_products_{current_time}.csv"
         product_feed = self._create_product_feed(file_name, catalog)
@@ -150,9 +147,11 @@ class VtexService:
             products_csv,
             file_name,
         )
+        print("Uploading the CSV file to Facebook completed successfully")
         return product_feed
 
     def _create_product_feed(self, name, catalog: Catalog) -> ProductFeed:
+        print("Creating the product feed")
         service = self.fb_service(catalog.app)
         response = service.create_product_feed(catalog.facebook_catalog_id, name)
 
@@ -170,6 +169,7 @@ class VtexService:
     def _upload_product_feed(
         self, app, product_feed_id, csv_file, file_name, update_only=False
     ):
+        print("Uploading the product feed to facebook")
         service = self.fb_service(app)
         response = service.upload_product_feed(
             product_feed_id, csv_file, file_name, "text/csv", update_only
@@ -179,15 +179,212 @@ class VtexService:
 
         return True
 
-    def _update_products_on_facebook(
-        self, products_csv, catalog: Catalog, product_feed
-    ) -> bool:
+
+class ProductUpdateService(VtexServiceBase):
+    def __init__(
+        self,
+        api_credentials: APICredentials,
+        catalog: Catalog,
+        skus_ids: list,
+        product_feed: ProductFeed,
+    ):
+        super().__init__()
+        self.api_credentials = api_credentials
+        self.catalog = catalog
+        self.skus_ids = skus_ids
+        self.product_feed = product_feed
+        self.app = self.catalog.app
+        self.feed_id = self.product_feed.facebook_feed_id
+        self.fba_service = self.fb_service(self.app)
+        self.redis = get_redis_connection()
+        self.last_update_id = None
+
+    def webhook_product_insert(self):
+        pvt_service = self.get_private_service(
+            self.api_credentials.app_key, self.api_credentials.app_token
+        )
+        products_dto = pvt_service.update_webhook_product_info(
+            self.api_credentials.domain, self.skus_ids, self.catalog.vtex_app.config
+        )
+        if not products_dto:
+            return None
+
+        products_csv = pvt_service.data_processor.products_to_csv(products_dto)
+        update_successful = self._webhook_update_products_on_facebook(products_csv)
+
+        if update_successful is not True:
+            print("Not upload products on '_webhook_update_products_on_facebook'")
+            return None
+
+        # Removed on 03-30-2024
+        # self.product_manager.create_or_update_products_on_database(
+        #     products_dto, self.catalog, self.product_feed
+        # )
+        # products_list = pvt_service.data_processor.convert_dtos_to_dicts_list(
+        #     products_dto
+        # )
+        return products_dto
+
+    def _webhook_update_products_on_facebook(self, products_csv) -> bool:
         current_time = datetime.now().strftime("%Y-%m-%d_%H-%M")
-        file_name = f"update_{current_time}_{product_feed.name}"
-        return self._upload_product_feed(
-            catalog.app,
-            product_feed.facebook_feed_id,
-            products_csv,
-            file_name,
-            update_only=True,
+        file_name = f"update_{current_time}_{self.product_feed.name}"
+        upload_id = self._webhook_upload_product_feed(
+            csv_file=products_csv,
+            file_name=file_name,
+        )
+        upload_complete = self._wait_for_upload_completion(upload_id=upload_id)
+        if not upload_complete:
+            print("Upload did not complete within the expected time frame.")
+            return False
+
+        print("Finish update products to facebook")
+        return True
+
+    def _webhook_upload_product_feed(self, csv_file, file_name: str) -> str:
+        print("Uploading the product feed to facebook")
+        update_only = True
+        response = self.fba_service.upload_product_feed(
+            self.feed_id, csv_file, file_name, "text/csv", update_only
+        )
+        if "id" not in response:
+            raise FileNotSendValidationError()
+        return response["id"]
+
+    def _wait_for_upload_completion(self, upload_id) -> bool:
+        wait_time = 5
+        max_wait_time = 20 * 60
+        total_wait_time = 0
+
+        while total_wait_time < max_wait_time:
+            upload_complete = self.fba_service.get_upload_status_by_feed(
+                self.feed_id, upload_id
+            )
+            if upload_complete:
+                return True
+
+            print(
+                f"Waiting {wait_time} seconds to get feed: {self.feed_id} upload {upload_id} status."
+            )
+            time.sleep(wait_time)
+            total_wait_time += wait_time
+            wait_time = min(wait_time * 2, 160)
+
+        return False
+
+
+class CatalogProductInsertion:
+    @classmethod
+    def first_product_insert_with_catalog(cls, vtex_app: App, catalog_id: str):
+        """Inserts the first product with the given catalog."""
+        wpp_cloud_uuid = cls._get_wpp_cloud_uuid(vtex_app)
+        credentials = cls._get_credentials(vtex_app)
+        wpp_cloud = cls._get_wpp_cloud(wpp_cloud_uuid)
+
+        catalog = cls._get_or_sync_catalog(wpp_cloud, catalog_id)
+        cls._delete_existing_feeds_ifexists(catalog)
+        cls._update_app_connected_catalog_flag(wpp_cloud)
+        cls._link_catalog_to_vtex_app_if_needed(catalog, vtex_app)
+
+        cls._send_insert_task(credentials, catalog)
+
+    @staticmethod
+    def _get_wpp_cloud_uuid(vtex_app) -> str:
+        """Retrieves WPP Cloud UUID from VTEX app config."""
+        wpp_cloud_uuid = vtex_app.config.get("wpp_cloud_uuid")
+        if not wpp_cloud_uuid:
+            raise ValueError(
+                "The VTEX app does not have the WPP Cloud UUID in its configuration."
+            )
+        return wpp_cloud_uuid
+
+    @staticmethod
+    def _get_credentials(vtex_app) -> dict:
+        """Extracts API credentials from VTEX app config."""
+        api_credentials = vtex_app.config.get("api_credentials", {})
+        if not all(
+            key in api_credentials for key in ["app_key", "app_token", "domain"]
+        ):
+            raise ValueError("Missing one or more API credentials.")
+        return api_credentials
+
+    @staticmethod
+    def _get_wpp_cloud(wpp_cloud_uuid) -> App:
+        """Fetches the WPP Cloud app based on UUID."""
+        try:
+            return App.objects.get(uuid=wpp_cloud_uuid)
+        except App.DoesNotExist:
+            raise ValueError(
+                f"The cloud app {wpp_cloud_uuid} linked to the VTEX app does not exist."
+            )
+
+    @classmethod
+    def _get_or_sync_catalog(cls, wpp_cloud, catalog_id) -> Catalog:
+        from marketplace.wpp_products.tasks import FacebookCatalogSyncService
+
+        """Attempts to find the catalog, syncs if not found, and tries again."""
+        catalog = wpp_cloud.catalogs.filter(facebook_catalog_id=catalog_id).first()
+        if not catalog:
+            print(
+                f"Catalog {catalog_id} not found for cloud app: {wpp_cloud.uuid}. Starting catalog synchronization."
+            )
+            sync_service = FacebookCatalogSyncService(wpp_cloud)
+            sync_service.sync_catalogs()
+            catalog = wpp_cloud.catalogs.filter(facebook_catalog_id=catalog_id).first()
+            if not catalog:
+                raise ValueError(
+                    f"Catalog {catalog_id} not found for cloud app: {wpp_cloud.uuid} after synchronization."
+                )
+        return catalog
+
+    @staticmethod
+    def _link_catalog_to_vtex_app_if_needed(catalog, vtex_app) -> None:
+        from django.contrib.auth import get_user_model
+
+        """Links the catalog to the VTEX app if not already linked."""
+        if not catalog.vtex_app:
+            User = get_user_model()
+            catalog.vtex_app = vtex_app
+            catalog.modified_by = User.objects.get_admin_user()
+            catalog.save()
+            print(
+                f"Catalog {catalog.name} successfully linked to VTEX app: {vtex_app.uuid}."
+            )
+
+    @staticmethod
+    def _delete_existing_feeds_ifexists(catalog) -> None:
+        """Deletes existing feeds linked to the catalog and logs their IDs."""
+        feeds = catalog.feeds.all()
+        total = feeds.count()
+        if total > 0:
+            print(f"Deleting {total} feed(s) linked to catalog {catalog.name}.")
+            for feed in feeds:
+                print(f"Deleting feed with ID {feed.facebook_feed_id}.")
+                feed.delete()
+            print(
+                f"All feeds linked to catalog {catalog.name} have been successfully deleted."
+            )
+        else:
+            print(f"No feeds linked to catalog {catalog.name} to delete.")
+
+    @staticmethod
+    def _update_app_connected_catalog_flag(app) -> None:
+        """Change connected catalog status"""
+        connected_catalog = app.config.get("connected_catalog", None)
+        if connected_catalog is not True:
+            app.config["connected_catalog"] = True
+            app.save()
+            print("Changed connected_catalog to True")
+
+    @staticmethod
+    def _send_insert_task(credentials, catalog) -> None:
+        from marketplace.celery import app as celery_app
+
+        """Sends the insert task to the task queue."""
+        celery_app.send_task(
+            name="task_insert_vtex_products",
+            kwargs={"credentials": credentials, "catalog_uuid": str(catalog.uuid)},
+            queue="product_first_synchronization",
+        )
+        print(
+            f"Catalog: {catalog.name} was sent successfully sent to task_insert_vtex_products"
         )