+++ marketplace/services/vtex/private/products/service.py
@@ -30,6 +30,7 @@ Example:
         products = service.list_all_products("domain.vtex.com")
         # Use products data as needed
 """
+
 from typing import List
 
 from marketplace.services.vtex.exceptions import CredentialsValidationError
@@ -39,9 +40,9 @@ from marketplace.services.vtex.utils.data_processor import FacebookProductDTO
 
 
 class PrivateProductsService:
-    def __init__(self, client, data_processor=DataProcessor):
+    def __init__(self, client, data_processor_class=DataProcessor):
         self.client = client
-        self.data_processor = data_processor
+        self.data_processor = data_processor_class()
         # TODO: Check if it makes sense to leave the domain instantiated
         # so that the domain parameter is removed from the methods
 
@@ -78,27 +79,20 @@ class PrivateProductsService:
         )  # TODO: Change to pvt_simulate_cart_for_seller
 
     def update_webhook_product_info(
-        self, domain, webhook_payload, config
+        self, domain: str, skus_ids: list, config: dict
     ) -> List[FacebookProductDTO]:
-        sku_id = webhook_payload["IdSku"]
-        price_modified = webhook_payload["PriceModified"]
-        stock_modified = webhook_payload["StockModified"]
-        other_changes = webhook_payload["HasStockKeepingUnitModified"]
-
         seller_ids = self.client.list_active_sellers(domain)
-
-        if price_modified or stock_modified or other_changes:
-            rules = self._load_rules(config.get("rules", []))
-            store_domain = config.get("store_domain")
-            updated_products_dto = self.data_processor.process_product_data(
-                [sku_id],
-                seller_ids,
-                self,
-                domain,
-                store_domain,
-                rules,
-                update_product=True,
-            )
+        rules = self._load_rules(config.get("rules", []))
+        store_domain = config.get("store_domain")
+        updated_products_dto = self.data_processor.process_product_data(
+            skus_ids,
+            seller_ids,
+            self,
+            domain,
+            store_domain,
+            rules,
+            update_product=True,
+        )
 
         return updated_products_dto